implemented token expiry check

+++ foodprint/lib/infrastructure/auth/custom_auth_facade.dart
@@ -1,5 +1,3 @@
-
-
 import 'package:flutter/foundation.dart';
 import 'package:dartz/dartz.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
@@ -18,24 +16,19 @@ import 'package:http/http.dart' as http;
 // Implements the interface authentication methods (communicates with the outside world)
 @LazySingleton(as: IAuthFacade)
 class CustomAuthFacade implements IAuthFacade {
-
   @override
-  Future<Either<RegisterFailure, Unit>> register({
-    @required EmailAddress emailAddress, 
-    @required Username username,
-    @required Password password
-  }) async {
+  Future<Either<RegisterFailure, Unit>> register(
+      {@required EmailAddress emailAddress,
+      @required Username username,
+      @required Password password}) async {
     final emailAddressStr = emailAddress.getOrCrash();
     final usernameStr = username.getOrCrash();
-    final passwordStr = password.getOrCrash(); 
-    final res = await http.post(
-      "$serverIP/api/users/register",
-      body: {
-        "email": emailAddressStr,
-        "username": usernameStr,
-        "password": passwordStr
-      }
-    );
+    final passwordStr = password.getOrCrash();
+    final res = await http.post("$serverIP/api/users/register", body: {
+      "email": emailAddressStr,
+      "username": usernameStr,
+      "password": passwordStr
+    });
     if (res.statusCode == 200) {
       return right(unit);
     } else if (res.statusCode == 401) {
@@ -46,29 +39,22 @@ class CustomAuthFacade implements IAuthFacade {
   }
 
   @override
-  Future<Either<LoginFailure, JWT>> login({
-    @required Username username,
-    @required Password password
-  }) async {
+  Future<Either<LoginFailure, JWT>> login(
+      {@required Username username, @required Password password}) async {
     final usernameStr = username.getOrCrash();
-    final passwordStr = password.getOrCrash(); 
-    final res = await http.post(
-      "$serverIP/api/users/login", 
-      body: {
-        "username": usernameStr,
-        "password": passwordStr
-      }
-    );
+    final passwordStr = password.getOrCrash();
+    final res = await http.post("$serverIP/api/users/login",
+        body: {"username": usernameStr, "password": passwordStr});
     if (res.statusCode == 200) {
-       
-      final JWT jwt = JWT(token: res.body); // token should be valid but check in case of tampering
+      final JWT jwt = JWT(
+          token:
+              res.body); // token should be valid but check in case of tampering
       if (jwt.isValid()) {
-
-        // Store token 
-        final localDataClient = JWTStorageClient(storage: const FlutterSecureStorage());
+        // Store token
+        final localDataClient =
+            JWTStorageClient(storage: const FlutterSecureStorage());
         await localDataClient.storeUserToken(jwt);
         return right(jwt);
-        
       } else {
         return left(const LoginFailure.serverError());
       }
@@ -81,27 +67,38 @@ class CustomAuthFacade implements IAuthFacade {
 
   @override
   Future<Option<JWT>> getUserToken() async {
-    final localDataClient = JWTStorageClient(storage: const FlutterSecureStorage()); 
+    final localDataClient =
+        JWTStorageClient(storage: const FlutterSecureStorage());
     try {
-      final jwt = await localDataClient.getUserToken(); // try to fetch the jwt
-      return some(jwt);
+      final JWT jwt =
+          await localDataClient.getUserToken(); // try to fetch token
+
+      // Check if token expired
+      final String token = jwt.getOrCrash();
+      Map<String, dynamic> payload = JWT.getDecodedPayload(token);
+
+      final DateTime expiry =
+          DateTime.fromMillisecondsSinceEpoch((payload["exp"] as int) * 1000);
+
+      if (expiry.isAfter(DateTime.now())) { // hasn't expired
+        return some(jwt);
+      }
+
+      // If token expired, delete and return none()
+      await localDataClient.deleteUserToken();
+      return none();
     } on TokenException {
       return none();
-    } 
+    }
   }
 
   @override
   Future<void> logout({@required User user}) async {
-
-    final localDataClient = JWTStorageClient(storage: const FlutterSecureStorage()); 
+    final localDataClient =
+        JWTStorageClient(storage: const FlutterSecureStorage());
     final username = user.username.getOrCrash();
 
-    await http.post(
-        "$serverIP/api/users/logout",
-        body: {
-          "username": username
-        }
-    );
-    await localDataClient.deleteUserToken(); 
+    await http.post("$serverIP/api/users/logout", body: {"username": username});
+    await localDataClient.deleteUserToken();
   }
 }
